requirements.txt
flask
flask-session
werkzeug
python-dotenv
google-generativeai
pdfplumber
python-docx
python-pptx
pytesseract
Pillow